--- app/src/main/java/com/eveggies/tujijenge/navigation.kt
@@ -1,2 +1,0 @@
-package com.eveggies.tujijenge
-
